backend/requirements.txt
Flask
Flask-Cors
python-dotenv
psycopg2-binary
requests